+++ migrations/sqlite/20251112180821_extend_dataset_statistics_with_object_links.sql
@@ -1,0 +1,8 @@
+/* ------------------------------ */
+
+ALTER TABLE dataset_statistics ADD COLUMN num_object_links BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE dataset_statistics ADD COLUMN object_links_size BIGINT NOT NULL DEFAULT 0;
+
+DELETE FROM dataset_statistics;
+
+/* ------------------------------ */