+++ src/infra/datasets/postgres/src/repos/postgres_dataset_statistics_repository.rs
@@ -49,7 +49,13 @@ impl DatasetStatisticsRepository for PostgresDatasetStatisticsRepository {
 
         let maybe_record = sqlx::query!(
             r#"
-            SELECT last_pulled, num_records, data_size, checkpoints_size FROM dataset_statistics
+            SELECT  last_pulled,
+                    num_records,
+                    data_size,
+                    checkpoints_size,
+                    num_object_links,
+                    object_links_size
+            FROM dataset_statistics
                 WHERE dataset_id = $1 AND block_ref_name = $2
             "#,
             stack_dataset_id.as_str(),
@@ -65,6 +71,8 @@ impl DatasetStatisticsRepository for PostgresDatasetStatisticsRepository {
                 num_records: u64::try_from(record.num_records).unwrap(),
                 data_size: u64::try_from(record.data_size).unwrap(),
                 checkpoints_size: u64::try_from(record.checkpoints_size).unwrap(),
+                num_object_links: u64::try_from(record.num_object_links).unwrap(),
+                object_links_size: u64::try_from(record.object_links_size).unwrap(),
             })
         } else {
             Err(DatasetStatisticsNotFoundError {
@@ -73,6 +81,40 @@ impl DatasetStatisticsRepository for PostgresDatasetStatisticsRepository {
             }
             .into())
         }
+    }
+
+    async fn get_total_statistic_by_account_id(
+        &self,
+        account_id: &odf::AccountID,
+    ) -> Result<TotalStatistic, InternalError> {
+        let mut tr = self.transaction.lock().await;
+        let connection_mut = tr.connection_mut().await?;
+
+        let record = sqlx::query!(
+            r#"
+                SELECT
+                    (COALESCE(SUM(ds.num_records), 0))::BIGINT         AS num_records,
+                    (COALESCE(SUM(ds.data_size), 0))::BIGINT           AS data_size,
+                    (COALESCE(SUM(ds.checkpoints_size), 0))::BIGINT    AS checkpoints_size,
+                    (COALESCE(SUM(ds.num_object_links), 0))::BIGINT    AS num_object_links,
+                    (COALESCE(SUM(ds.object_links_size), 0))::BIGINT   AS object_links_size
+                FROM dataset_statistics ds
+                    LEFT JOIN dataset_entries de ON ds.dataset_id = de.dataset_id
+                WHERE de.owner_id = $1
+            "#,
+            account_id.to_string()
+        )
+        .fetch_one(connection_mut)
+        .await
+        .int_err()?;
+
+        Ok(TotalStatistic {
+            num_records: u64::try_from(record.num_records.unwrap()).unwrap(),
+            data_size: u64::try_from(record.data_size.unwrap()).unwrap(),
+            checkpoints_size: u64::try_from(record.checkpoints_size.unwrap()).unwrap(),
+            num_object_links: u64::try_from(record.num_object_links.unwrap()).unwrap(),
+            object_links_size: u64::try_from(record.object_links_size.unwrap()).unwrap(),
+        })
     }
 
     async fn set_dataset_statistics(
@@ -86,14 +128,25 @@ impl DatasetStatisticsRepository for PostgresDatasetStatisticsRepository {
 
         sqlx::query!(
             r#"
-            INSERT INTO dataset_statistics (dataset_id, block_ref_name, last_pulled, num_records, data_size, checkpoints_size)
-                VALUES ($1, $2, $3, $4, $5, $6)
+            INSERT INTO dataset_statistics (
+                dataset_id,
+                block_ref_name,
+                last_pulled,
+                num_records,
+                data_size,
+                checkpoints_size,
+                num_object_links,
+                object_links_size
+            )
+                VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
                 ON CONFLICT (dataset_id, block_ref_name)
                 DO UPDATE SET
                     last_pulled = EXCLUDED.last_pulled,
                     num_records = EXCLUDED.num_records,
                     data_size = EXCLUDED.data_size,
-                    checkpoints_size = EXCLUDED.checkpoints_size
+                    checkpoints_size = EXCLUDED.checkpoints_size,
+                    num_object_links = EXCLUDED.num_object_links,
+                    object_links_size = EXCLUDED.object_links_size
             "#,
             dataset_id.to_string(),
             block_ref.as_str(),
@@ -101,6 +154,8 @@ impl DatasetStatisticsRepository for PostgresDatasetStatisticsRepository {
             i64::try_from(statistics.num_records).unwrap(),
             i64::try_from(statistics.data_size).unwrap(),
             i64::try_from(statistics.checkpoints_size).unwrap(),
+            i64::try_from(statistics.num_object_links).unwrap(),
+            i64::try_from(statistics.object_links_size).unwrap(),
         )
         .execute(connection_mut)
         .await